accessing user data
accessing the data that the user inputed about their farm. called it from the state and allows the app from the create farm page on to function and continue to be developed.
works on chrome and edge to call data (paddocks and herds) from the farm that had just been created.

next steps include:
saving the farm that has been created to the user id. This means that when the user logs in and out their farm is saved.

creating a placebo page that will be filled in with the specific paddock/ herd that the user chose. Figuring out how to actually call the data based on what the user chose so that it chnages everytime they chose something else. the page should be versitle.